🚚 Rename :  type 폴더를 shared 밑으로 내림
