+++ src/features/calendar/dotCalendar/components/MultiDot.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { View } from 'react-native';
 
-import { CalendarSchedule } from '../../types';
+import { CalendarSchedule } from '../../shared/types';
 
 interface MultiDotProps {
   schedules: CalendarSchedule[];
