+++ src/features/calendar/shared/api/calendarApi.ts
@@ -1,4 +1,4 @@
-import { CalendarScheduleRequest, CalendarScheduleResponse } from '../../types';
+import { CalendarScheduleRequest, CalendarScheduleResponse } from '../types';
 
 import { axiosInstance } from '@/shared/api/axiosInstance';
